# docs/openapi.json
{
  "components": {
    "schemas": {
      "ApiError": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "additionalProperties": false,
        "properties": {
          "code": {
            "type": "string"
          },
          "hint": {
            "type": "string"
          },
          "issues": {
            "items": {
              "additionalProperties": false,
              "properties": {
                "message": {
                  "type": "string"
                },
                "path": {
                  "type": "string"
                }
              },
              "required": [
                "path",
                "message"
              ],
              "type": "object"
            },
            "type": "array"
          },
          "message": {
            "type": "string"
          },
          "request_id": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "type": "object"
      }
    },
    "securitySchemes": {
      "accessAssertion": {
        "description": "The signed Access JWT. Set by Cloudflare Access at the edge; the Worker verifies it itself rather than trusting that Access ran. Browsers may instead present the `CF_Authorization` cookie, which Cloudflare documents as not guaranteed to be passed in every context — the header is primary.",
        "in": "header",
        "name": "Cf-Access-Jwt-Assertion",
        "type": "apiKey"
      },
      "accessClientId": {
        "description": "An Access **service token** id. Presented with `CF-Access-Client-Secret`; Access exchanges the pair at the edge, so this header never reaches the Worker.",
        "in": "header",
        "name": "CF-Access-Client-Id",
        "type": "apiKey"
      },
      "accessClientSecret": {
        "description": "The matching Access service token secret. Required together with `CF-Access-Client-Id`.",
        "in": "header",
        "name": "CF-Access-Client-Secret",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "A self-hosted secrets manager on Cloudflare Workers and D1.\n\n**Slug alias routes.** Every environment-scoped route is served at two paths:\nthe canonical `/projects/{project}/environments/{env}/…` and the shorter\n`/p/{project}/e/{env}/…` for CLI ergonomics. Only the canonical form appears\nbelow; the alias is byte-identical in behaviour, matches **exactly** and never\nas a prefix.\n\n**There is no CORS.** This API emits no `Access-Control-Allow-Origin` and never\nwill. Omitting it is what stops another site reading a response in a victim's\nbrowser. The admin UI is same-origin and needs nothing.\n\n**`404` is used for absent and invisible alike.** A resource you cannot see is\nreported exactly as one that does not exist. The alternative leaks which project\nnames are in use to a caller with no grants at all.\n\n**Validation never echoes input.** A `422` body carries `issues` of\n`{path, message}`. The rejected value is dropped before the body is built, so a\nfailed secret write cannot put a plaintext value in a response, a log line or an\naudit row.",
    "title": "prick",
    "version": "0.0.0-dev"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/projects/{project}/environments/{env}": {
      "delete": {
        "description": "Requires **admin** at the environment scope. One statement; the cascade is the database's. The audit row records how many secrets went with it.",
        "operationId": "deleteEnvironment",
        "parameters": [
          {
            "in": "path",
            "name": "project",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "env",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "`BAD_REQUEST` — a malformed `If-Match`, or a precondition this API cannot express."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "`UNAUTHENTICATED` — no valid Cloudflare Access assertion was presented."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "`FORBIDDEN` — authenticated, but no grant covers this scope. The denial is audited."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "`NOT_FOUND` — absent **or** invisible, deliberately indistinguishable. Returning 403 for one and 404 for the other would turn this API into an oracle for which slugs are in use."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "`VALIDATION_FAILED` — schema rejection, including an unknown field. `issues` carries `{path, message}` and never the submitted value."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "`INTERNAL`, `DECRYPT_FAILED`, `UNKNOWN_KID` or `SERVER_MISCONFIGURED`."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "`NO_ADMINS_CONFIGURED` or `IDENTITY_PROVIDER_UNAVAILABLE`."
          }
        },
        "summary": "Delete an environment",
        "tags": [
          "environments"
        ]
      },
      "get": {
        "description": "Carries `rev`, the optimistic-concurrency token. It is the same number the secret collection returns as its `ETag`.",
        "operationId": "getEnvironment",
        "parameters": [
          {
            "in": "path",
            "name": "project",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "env",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "additionalProperties": false,
                  "properties": {
                    "description": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "id": {
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "projectId": {
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
                      "type": "string"
                    },
                    "rev": {
                      "maximum": 9007199254740991,
                      "minimum": 0,
                      "type": "integer"
                    },
                    "secretCount": {
                      "maximum": 9007199254740991,
                      "minimum": 0,
                      "type": "integer"
                    },
                    "slug": {
                      "maxLength": 64,
                      "minLength": 1,
                      "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$",
                      "type": "string"
                    },
                    "updatedAt": {
                      "maximum": 9007199254740991,
                      "minimum": 0,
                      "type": "integer"
                    }
                  },
                  "required": [
                    "id",
                    "projectId",
                    "slug",
                    "name",
                    "description",
                    "rev",
                    "secretCount",
                    "updatedAt"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "The environment."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "`BAD_REQUEST` — a malformed `If-Match`, or a precondition this API cannot express."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "`UNAUTHENTICATED` — no valid Cloudflare Access assertion was presented."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "`FORBIDDEN` — authenticated, but no grant covers this scope. The denial is audited."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "`NOT_FOUND` — absent **or** invisible, deliberately indistinguishable. Returning 403 for one and 404 for the other would turn this API into an oracle for which slugs are in use."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "`VALIDATION_FAILED` — schema rejection, including an unknown field. `issues` carries `{path, message}` and never the submitted value."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "`INTERNAL`, `DECRYPT_FAILED`, `UNKNOWN_KID` or `SERVER_MISCONFIGURED`."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "`NO_ADMINS_CONFIGURED` or `IDENTITY_PROVIDER_UNAVAILABLE`."
          }
        },
        "summary": "Get an environment",
        "tags": [
          "environments"
        ]
      }
    },
    "/api/v1/projects/{project}/environments/{env}/secrets": {
      "get": {
        "description": "Key names, versions and metadata. **No values, ever.** Every row is nonetheless decrypted and the plaintext discarded immediately, because `unreadable` cannot be determined any other way — AES-GCM has no verify-without-decrypting operation, the tag check IS the decryption. A row that fails comes back marked rather than omitted, and the read is audited with `outcome: 'error'`.",
        "operationId": "listSecrets",
        "parameters": [
          {
            "in": "path",
            "name": "project",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "env",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "items": {
                    "additionalProperties": false,
                    "properties": {
                      "description": {
                        "anyOf": [
                          {
                            "type": "string"
                          },
                          {
                            "type": "null"
                          }
                        ]
                      },
                      "key": {
                        "maxLength": 256,
                        "minLength": 1,
                        "pattern": "^[A-Za-z_][A-Za-z0-9_]*$",
                        "type": "string"
                      },
                      "kid": {
                        "anyOf": [
                          {
                            "type": "string"
                          },
                          {
                            "type": "null"
                          }
                        ]
                      },
                      "unreadable": {
                        "type": "boolean"
                      },
                      "updatedAt": {
                        "maximum": 9007199254740991,
                        "minimum": 0,
                        "type": "integer"
                      },
                      "updatedBy": {
                        "type": "string"
                      },
                      "version": {
                        "exclusiveMinimum": 0,
                        "maximum": 9007199254740991,
                        "type": "integer"
                      }
                    },
                    "required": [
                      "key",
                      "description",
                      "version",
                      "updatedAt",
                      "updatedBy",
                      "kid",
                      "unreadable"
                    ],
                    "type": "object"
                  },
                  "type": "array"
                }
              }
            },
            "description": "The secrets, ordered by key.",
            "headers": {
              "ETag": {
                "description": "The environment's revision, as a strong entity tag. Send it back as `If-Match` on a write to make that write conditional.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "`BAD_REQUEST` — a malformed `If-Match`, or a precondition this API cannot express."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "`UNAUTHENTICATED` — no valid Cloudflare Access assertion was presented."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "`FORBIDDEN` — authenticated, but no grant covers this scope. The denial is audited."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "`NOT_FOUND` — absent **or** invisible, deliberately indistinguishable. Returning 403 for one and 404 for the other would turn this API into an oracle for which slugs are in use."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "`VALIDATION_FAILED` — schema rejection, including an unknown field. `issues` carries `{path, message}` and never the submitted value."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "`INTERNAL`, `DECRYPT_FAILED`, `UNKNOWN_KID` or `SERVER_MISCONFIGURED`."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "`NO_ADMINS_CONFIGURED` or `IDENTITY_PROVIDER_UNAVAILABLE`."
          }
        },
        "summary": "List an environment's secrets",
        "tags": [
          "secrets"
        ]
      }
    },
...
  },
  "security": [
    {
      "accessClientId": [],
      "accessClientSecret": []
    },
    {
      "accessAssertion": []
    }
  ],
  "servers": [
    {
      "description": "This deployment.",
      "url": "/api/v1"
    }
  ],
...
}